refactor part 1, break up some small functions
